warning fix
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@4743 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ base/XmlExportable.C
@@ -34,7 +34,7 @@ std::string XmlExportable::encode(const std::string &s0)
     size_t buflen = 0;
 
     static char multibyte[20];
-    int mblen = 0;
+    size_t mblen = 0;
 
     size_t len = s0.length();
